Fix user_fixture and add auth test that uses it
The fixture was returning the result of `Accounts.register_user/1`,
which is a 2 element tuple (`{ :ok, user }` in the success case) but I
was trying to use the code as if it simply returned a `User` struct.

Erlang was giving me what was actually a rather nice error message, in
that it was printing out the tuple and telling me it wasn't a `map` or a
module and so it couldn't `apply` using the dot syntax, but I was too
distracted by the fact that all of the `User` key/value pairs were
visible in the output to see that they were nested in a tuple.

Lesson: read error messages closely :) even if it means extracting them
out to an editor to parse/ format them