Add configuration to limit hashing in test environment
Since we are going through the complete pipeline in our tests, including
those that require users (and thus their registration in the
`user_fixture`), we are running into poor time scaling as we add more
tests.

This is due to the hashing being performed on passwords for our mock
users by `pbkdf2_elixir`. This configuration limits the hashing to a
single round which vastly reduces the time spent on it.

We do not need a "production-like" level of password security for our
tests, and we want the tests to run as fast as possible to encourage us
to run them frequently.