Persist annotations to the database
Annotations are now held onto in the database, with a couple of
functions added to the `Multimedia` context for interacting with them.

A side-effect of this is that we can now see who posts what annotations
by way of their username :)

I also had to modify a test slightly because I reset the development
database and forgot that the `seeds.exs` file populates the `categories`
table. To keep the test clear and not reliant on the outside data I
added a filter to keep the assertion focused only on the categories
created by the test.