Add "presence" to see who is watching what
For how powerful this feature is, I think it's worth noting that this
took very little effort with the rest of the channel and rendering code
already in place.

You more-or-less can just bolt it on, define a fetch callback to get the
data that your client will care about, and away you go.