Update ecto
I had a problem running `:observer.start` in iex, which ships with
erlang, so after some googling I ended up chasing some homebrew stuff
down and basically just installing fresh, latest builds for everything.

The problem appears to be an issue with macOS Big Sur and the widget
library that the observer is rendered with.

After doing all this and trying to re-compile my app, Ecto was failing
to compile. One of the suggestions in the output is to try updating it,
and that worked.

It seems that the `Ecto.Type` behaviour now requires a few new
callbacks, which I do not have implemented for my `Permalink`, but I
will get to that soon.