Reorganize app as an "umbrella" project
This allows the separate "apps", in the OTP sense, to be truly separate.

The web code is now its own app, with its own children. Likewise, the
data side of the application (containing the `Accounts`, `Videos`, and `Multimedia` contexts) is its own app, with its own children (like Ecto).

This moves me closer to having a more "self-healing" architecture, with
individual pieces able to fail and restart according to their
supervision trees. It also enforces the separation of the data and web
code.

All of the "apps" use the same dependencies, defined in the `mix.exs`
file in the root directory.