Add a cache for info systems
The cache is based on the `:ets` or `Erlang Term Storage` application
that ships with `OTP`.
It sweeps itself every 60 seconds (by default) so that its memory
footprint does not grow indefinitely. This is perfectly fine for our use
case, since we just want to keep around questions long enough to imagine
that a user won't ask them again